Readme in code overview
The readme files are now rendered in the code overview below
the file tree and not in a separate page. If the user changes the
directory or the revision, then the readme file of that revision or
directory gets rendered, if it exists.

This change depends on core changes in release 3.1.0.