Fix the old readme url
The readme view was moved into an extension point of the code overview. Therefore the old url broke. This fixes the old url, which is now redirected to the code overview with the readme anchor. Additionally the alwaysMatches function for the ReadmeNavLink was replaced, by a funtion always returning false. Since the readme and the code overview now share the same site, the readme link is never supposed to be active.